revert: revert folder structure for review
